+++ docs/changelog.rst
@@ -19,6 +19,7 @@ Changelog
 * Tools: it is now possible to unset the current object in reference properties by pressing the Delete key.
 * Tools: added drag & drop support to the Level Tree and other object trees.
 * Tools: added the ability to import external textures from FBX scenes.
+* Tools: added an option to calculate mesh tangents when importing scenes.
 * Runtime: added the ability to override render settings in :doc:`boot.config <reference/boot_config>`.
 * Runtime: added the ability to specify a configuration file with user-specific settings via the ``user_config`` setting in :doc:`boot.config <reference/boot_config>`.
 * Runtime: Linux: it is now possible to run with either OpenGL or Vulkan rendering backends.